slkm: change directory structure
Change the directory structure to make it easier to build the core
safe C library and then package it up for a Linux kernel module (slkm).

Also, added some documentation related to the newly ratified C11
standard and a script to identify potential replacements within
a code base, see 'scripts/'.