Bugfix: compile error missing type
== Problem
When the library was installed to the system and a user attempted
to compile against the library the compile would fail.

== Reason
The problem was due to some types not being defined in the public API
of the library.

== Solution
Changed the public API, header files, of the library so that a few
of the files are now auto-generated based on information discovered
by the configure script. The header files that are now auto-generated
are 'safe_lib_errno.h' and 'safe_types.h'.