add wcsfc_s, full case-folding of string to lower
extends dest with multiple fold (i.e. mostly lower) case towfc_s expansions,
add all the SpecialCasing rules from Unicode 10.0:
final greek sigma, lithuanian I and J-dot, turkish and azeri I-dot.
in some 1-char cases fc is different to lc.

Unicode 10.0 has 297 code points for which the fold case mapping is not to itself,
or to tolower(), 194 of it return length 1, handled in _towfc_single(),
the rest 102 map to 2 or 3 and are handled in towfc().

fc not yet normalization to NFC.