start with wcsnorm_s, NFD normalization
The plan is to do canonical NFC normalization,
but currently we are only at NFD (as Apple).
We need to fix the perl generators for the tables to emit
multibyte utf-32, not just utf-8, and a proper
dec_rest table. (taken from and generated with cperl).
The mark table needs -fbracket-limit=512.
reorder and compose are not yet implemented.

also let wcsfc_s use it, the codepoint-specific
decomposition only so far.