fix run-time constraints (actions before)
c11 p.584:
K.3.1.4 Runtime-constraint violations

"If the runtime-constraints section for a function states an action to be
performed when a runtime-constraint violation occurs, the function shall
perform the action before calling the runtime-constraint handler.  If the
runtime-constraints section lists actions that are prohibited when a
runtime-constraint violation occurs, then such actions are prohibited to the
function both before calling the handler and after the handler returns.

The runtime-constraint handler might not return.  If the handler does
return, the library function whose runtime-constraint was violated shall
return some indication of failure as given by the returns section in the
function’s specification."