change mem{cpy,move,set}_s with smax/n=0 WIP
On C11 allow a zero size for the source, otherwise keep the old behavior
and fail with ESZEROL.  Note that this is a compile-time behavior.
You need to use a -std=c11 compiler for the library to get the wanted C11 Annex K
behavior.
Only the usage of the library with a std=c11 compiler is not enough.

See #27

TODO: HAVE_C11 check features, not just __STDC_WANT_LIB_EXT1__.
-std=c11 should enable it also.