windows: --disable-shared sets DISABLE_DLLIMPORT
and skips the dllexport/dllimport attributes.
We can now build static-only on windows, which eases debugging

libtool sets the pic_flag -DDLL_EXPORT when building a shared lib on
windows, so we could have used that instead. But we need it later in test
executables to see if -lsafec is used at all, or overridden by -lc
dynamically when using shared libs. So we need it in the runtime config.
And with a different name.