add debugging code
why bos is not effective inside the lib anymore,
only at the caller site, the test.

not with clang-7, nor gcc-6.
all dest sizes are proper at the test, and -1 inside the lib.
so we cannot use proper overflow checks at run-time, only at compile-time,
i.e. clang-only. and we cannot replace the superficial run-time checks
in the lib, when we cannot guarantee that the compile-time checks catch
them all.