Add --disable-constraint-handler option
undef the run-time invoke_safe_{str,mem}_constraint_handler function calls
in safe_config.h to avoid the large errormsg strings.  As in freebsd.

Most of the OS/libc folks complain about it, but it is useful, similar to
the malloc hooks.  They are only used in the error cases.

Closes #50