Fix *scanf_s documentation. #59
That's not yet implemented.  It will be huge vsscanf implementation (at
least the size of the current library), and will open a can of worms.  If
there's a nice vsscanf MIT code, I'll take it.  I only saw C++ code from
Android bionic though.  Maybe the FreeBSD variant.

We only do argument checks and pass it to the native vsscanf() libc
variants.