Set SAFECLIB_HAVE_C99 in safe_config.h
The SAFECLIB_HAVE_C99 preprocessor check is used in public header
and local C files. When an application uses the header files the
SAFECLIB_HAVE_C99 option has to be set to the same option used
when compiling the safeclib. Add this to the configure command.

Without this patch the safeclib was compiled with SAFECLIB_HAVE_C99
and exported the symbol _sprintf_s_chk, but my application used the
header and did not manually set SAFECLIB_HAVE_C99 and then used the
sprintf_s symbol, which is not exported by the shared safeclib.

Signed-off-by: Hauke Mehrtens <hauke.mehrtens@intel.com>